refactor: applied SOLID to gps repo
