+++ internal/repository/gps/gps_repository.go
@@ -2,7 +2,6 @@ package gps
 
 import (
 	"errors"
-	"fmt"
 	"v3/internal/domain"
 
 	"gorm.io/gorm"
@@ -26,26 +25,4 @@ type gpsRepository struct {
 
 func NewGPSRepository(db *gorm.DB) GPSRepository {
 	return &gpsRepository{DB: db}
-}
-
-func (r *gpsRepository) Create(d *domain.GPS) (*domain.GPS, error) {
-
-	if d.DeviceID == "" {
-		return nil, ErrDeviceIDEmpty
-	}
-	if d.Timestamp.IsZero() {
-		return nil, ErrTimestampEmpty
-	}
-	if d.Latitude == 0 {
-		return nil, ErrLatitudeEmpty
-	}
-	if d.Longitude == 0 {
-		return nil, ErrLongitudeEmpty
-	}
-
-	if err := r.DB.Create(d).Error; err != nil {
-		return nil, fmt.Errorf("%w: %v", ErrCreateFailed, err)
-	}
-
-	return d, nil
 }
